requirements.txt
azure-functions==1.13.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.9.0
huggingface-hub==0.12.1
idna==3.4
joblib==1.2.0
keybert==0.7.0
markdown-it-py==2.2.0
mdurl==0.1.2
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
Pillow==9.4.0
Pygments==2.14.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.2
scikit-learn==1.2.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
